refactor: http and chat clients with provider, clean up clientButton and properly separate clients, clean up layout
