+++ src/features/httpclients/components/RawResponseViewer.tsx
@@ -4,7 +4,7 @@
 
 import { Typography, styled } from '@mui/material';
 import { Labeled, RecordContextProvider, useTranslate } from 'react-admin';
-import { HttpClientResponse } from '../HttpClientProvider';
+import { HttpClientResponse } from '../provider/HttpClientProvider';
 import { StatusBadge } from '../HttpClient';
 
 export const RawResponseViewer = ({
